Remove AbstractReportModelHandler.java
